minor change to test gps process

+++ CANdaemon/src/gps/test_process/gps-dbus-server.h
@@ -7,7 +7,7 @@ enum {
 
 
 struct gps_status {
-    uint16_t current_state;
+    int32_t current_state;
 };
